requirements.txt
python-socketio[client]==5.10.0
picamera2
psutil==5.9.6
opencv-python==4.8.1.78
numpy==1.24.3
pillow
